requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
Pillow==10.0.0
qrcode==7.4.2
fpdf==1.7.2
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.2
reportlab==4.0.4